Introduce /unstable entrypoint, move stuff around
