+++ package.json
@@ -50,6 +50,18 @@
       },
       "default": "./dist/plugins/encoders/mp3.js"
     },
+    "./unstable": {
+      "types": "./dist/unstable/index.d.ts",
+      "node": {
+        "require": "./dist/unstable/index.node.cjs",
+        "import": "./dist/unstable/index.js"
+      },
+      "browser": {
+        "require": "./dist/unstable/index.browser.cjs",
+        "import": "./dist/unstable/index.js"
+      },
+      "default": "./dist/unstable/index.js"
+    },
     "./dist/css/*": {
       "default": "./dist/css/*"
     },
@@ -70,6 +82,9 @@
       ],
       "mp3-encoder": [
         "./dist/plugins/encoders/mp3.d.ts"
+      ],
+      "unstable": [
+        "./dist/unstable/index.d.ts"
       ]
     }
   },
