+++ src/components/Message/MessageOptions.tsx
@@ -101,7 +101,7 @@ const UnMemoizedMessageOptions = <
 };
 
 /**
- * @deprecated Use MessageActions_UNSTABLE (`MessageActions` component) instead.
+ * @deprecated Consider moving to `MessageActions` from `stream-chat-react/unstable` instead.
  */
 export const MessageOptions = React.memo(
   UnMemoizedMessageOptions,
