+++ src/components/MessageActions/index.ts
@@ -1,4 +1,3 @@
 export * from './MessageActions';
 export * from './MessageActionsBox';
 export * from './CustomMessageActionsList';
-export * as MessageActions_UNSTABLE from './MessageActions_UNSTABLE';
