+++ src/context/ComponentContext.tsx
@@ -102,7 +102,7 @@ export type ComponentContextValue<
   LoadingIndicator?: React.ComponentType<LoadingIndicatorProps>;
   /** Custom UI component to display a message in the standard `MessageList`, defaults to and accepts the same props as: [MessageSimple](https://github.com/GetStream/stream-chat-react/blob/master/src/components/Message/MessageSimple.tsx) */
   Message?: React.ComponentType<MessageUIComponentProps<StreamChatGenerics>>;
-  /** Custom UI component for message actions popup, accepts no props, all the defaults are set within [MessageActions_UNSTABLE](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MessageActions/MessageActions_UNSTABLE.tsx) (`MessageActions` component) */
+  /** Custom UI component for message actions popup, accepts no props, all the defaults are set within [MessageActions (unstable)](https://github.com/GetStream/stream-chat-react/blob/master/src/unstable/MessageActions/MessageActions.tsx) */
   MessageActions?: React.ComponentType;
   /** Custom UI component to display the contents of a bounced message modal. Usually it allows to retry, edit, or delete the message. Defaults to and accepts the same props as: [MessageBouncePrompt](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MessageBounce/MessageBouncePrompt.tsx) */
   MessageBouncePrompt?: React.ComponentType<MessageBouncePromptProps>;
